% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{view_dataframe}
\alias{view_dataframe}
\title{View Dataframe in RStudio Viewer}
\usage{
view_dataframe(data_name)
}
\arguments{
\item{data_name}{Character string with name of dataframe}
}
\value{
Character string with operation status
}
\description{
Opens a dataframe in the RStudio Viewer pane (read-only).
}
