% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{workspace_info}
\alias{workspace_info}
\title{Get Workspace Information}
\usage{
workspace_info()
}
\value{
Character string with workspace information
}
\description{
Returns information about the current R workspace (read-only).
}
