\name{zigguser-package}
\alias{zigguser-package}
\alias{zigguser}
\docType{package}
\title{\packageTitle{zigguser}}
\description{\packageDescription{zigguser}}
\section{Package Content}{\packageIndices{zigguser}}
\author{\packageAuthor{zigguser}}
\section{Maintainer}{\packageMaintainer{zigguser}}
\keyword{package}
