\name{zigguserDirect-package}
\alias{zigguserDirect-package}
\alias{zigguserDirect}
\docType{package}
\title{\packageTitle{zigguserDirect}}
\description{\packageDescription{zigguserDirect}}
\section{Package Content}{\packageIndices{zigguserDirect}}
\author{\packageAuthor{zigguserDirect}}
\section{Maintainer}{\packageMaintainer{zigguserDirect}}
\keyword{package}
